requirements.txt
absl-py==1.2.0
aiobotocore==2.3.4
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
aniso8601==9.0.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
botocore==1.24.21
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.29.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==1.12
frozenlist==1.3.1
fsspec==2022.7.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.2
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3fs==2022.7.1
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
